requirements.txt
anthropic==0.76.0
python-dotenv==1.2.1